Support zooming of VNC window
This is a feature which was often required by users who use notebooks
with a resolution which is than the resolution of the PalMA team monitor.

When zooming the VNC window, it will be centered on the screen, so it
is necessary to select the display layout after zooming to get
the correct layout again.

Signed-off-by: Stefan Weil <sw@weilnetz.de>