scripts: Fix screensaver running with dwb browser (invalid pin)
The default browser dwb restores previous browser windows by default.
If one of these windows is also a screensaver which calculates a new
pin, we get a race condition. The pin is calculated twice or more, and
the one which is shown in the screensaver display is not necessarily
the last and valid pin. In this case, users won't be able to enter a
correct pin.

When dwb is run with option --override-restore, it no longer restores
previous browser windows. This fixes the problem with invalid pins.

Signed-off-by: Stefan Weil <sw@weilnetz.de>