Launch impress in presentation mode
Added the --show parameter to launch slides in presentation mode by
default. This takes not just the full window, but the full screen!
Navigation through slides only works. when the file is not in window 1.
There seems to be a secound instance launched. Same problem seems to go
for calc files.